Added API request function, added and organized new types, minor styling, fixed clear form bug and notification bar display issue. Removed ARCHIVE/ finally.
