--- ARCHIVE/.jshintrc
@@ -1,3 +1,0 @@
-{
-    "esversion": 8
-}
