--- ARCHIVE/package.json
@@ -1,46 +1,0 @@
-{
-    "name": "lunch-memo",
-    "version": "0.1.0",
-    "description": "Lunch Memo makes it easy to organize lunch with your colleagues and other professionals in your area.",
-    "main": "index.js",
-    "directories": {
-        "test": "tests",
-        "src": "src"
-    },
-    "scripts": {
-        "start": "node ./bin/www",
-        "dev": "nodemon ./bin/www",
-        "sass:compile": "node-sass --include-path sass ./public/css/scss/main.scss ./public/css/main.css"
-    },
-    "proxy": "http://localhost:3000",
-    "author": "Bobby Wang & Andres Weber",
-    "license": "ISC",
-    "dependencies": {
-        "axios": "^0.21.1",
-        "bcryptjs": "^2.4.3",
-        "connect-ensure-login": "^0.1.1",
-        "connect-flash": "^0.1.1",
-        "cookie-parser": "^1.4.5",
-        "dotenv": "^8.2.0",
-        "express": "^4.17.1",
-        "express-flash": "0.0.2",
-        "express-session": "^1.17.1",
-        "hbs": "^4.1.1",
-        "heroku-ssl-redirect": "0.1.1",
-        "jsonwebtoken": "^8.5.1",
-        "mongoose": "^5.11.8",
-        "morgan": "^1.10.0",
-        "multer": "^1.4.2",
-        "node-sass-middleware": "^0.11.0",
-        "npm-sass": "^2.3.0",
-        "passport": "^0.4.1",
-        "passport-local": "^1.0.0",
-        "serve-favicon": "^2.5.0",
-        "sharp": "^0.26.3",
-        "validator": "^13.5.2",
-        "yelp-fusion": "^3.0.0"
-    },
-    "devDependencies": {
-        "nodemon": "^2.0.6"
-    }
-}
