--- ARCHIVE/public/browserconfig.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<browserconfig>
-    <msapplication>
-        <tile>
-            <square150x150logo src="/mstile-150x150.png"/>
-            <TileColor>#ff0000</TileColor>
-        </tile>
-    </msapplication>
-</browserconfig>
