--- ARCHIVE/public/css/scss/main.scss
@@ -1,24 +1,0 @@
-@import 'parts/settings';
-@import 'parts/auth';
-@import 'parts/normalize';
-@import 'parts/timepicker';
-@import 'parts/layout';
-@import 'parts/global';
-
-.bginfo {
-    position: absolute;
-    bottom: 2rem;
-    right: 2rem;
-    font-size: .8rem;
-    text-align: right;
-}
-
-.large-title {
-  font-size: 2.4rem;
-  font-weight: 300;
-  color: #fff;
-  text-shadow: 0 1px 1px rgba(0, 0, 0, 0.35);
-  margin-bottom: 1rem;
-  letter-spacing: 3px;
-  text-transform: uppercase;
-}
