--- ARCHIVE/public/css/scss/parts/settings.scss
@@ -1,18 +1,0 @@
-$primary-color: #7c5cbf;
-$gray: #777;
-$lightgray: #eee;
-$darkgray: #221e1f;
-$yellow: #f7be0d;
-$red: #d50000;
-$logoyellow: #ffe081;
-$header-font: 'Rubik', 'Open Sans', Helvetica, Arial, sans-serif;
-$body-font: 'Lato', 'Open Sans', Helvetica, Arial, sans-serif;
-$font-awesome: 'Font Awesome 5 Free';
-
-@mixin alpha-attribute($attribute, $color, $background) {
-    $percent: alpha($color) * 100%;
-    $opaque: opacify($color, 1);
-    $solid-color: mix($opaque, $background, $percent);
-    #{$attribute}: $solid-color;
-    #{$attribute}: $color;
-  }
