--- ARCHIVE/public/index.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]><!-->
-<html class="no-js">
-<!--<![endif]-->
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Lunch Memo</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel="apple-touch-icon" sizes="180x180" href="img/icon/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="img/icon/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="img/icon/favicon-16x16.png">
-    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#f5d808">
-    <meta name="apple-mobile-web-app-title" content="LUNCHmemo">
-    <meta name="application-name" content="LUNCHmemo">
-    <meta name="msapplication-TileColor" content="#ff0000">
-    <meta name="theme-color" content="#ffffff">
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
-
-    <link href="https://fonts.googleapis.com/css?family=Lato:400,700|Rubik:300,300i,700&display=swap" rel="stylesheet">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.7.2/animate.min.css" rel="stylesheet">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.10.2/css/all.min.css" rel="stylesheet">
-    <link href="css/main.css" rel="stylesheet">
-</head>
-
-<body>
-    <!--[if lt IE 7]>
-            <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="#">upgrade your browser</a> to improve your experience.</p>
-        <![endif]-->
-
-    <div class="app-wrap">
-        <header class="hide">
-            <h1 class="logo"><span class="lunch">LUNCH</span><span class="memo">memo</span></h1>
-            <nav>
-                <a href="#" class="mobile-menu"><i class="fas fa-bars"></i></a>
-                <div class="nav">
-                    Hello, <span id="username-nav"></span> <i class="fas fa-user-circle"></i><br />
-                    <a href="#" onclick="lmRunRegistration()">Logout</a>
-                </div>
-            </nav>
-        </header>
-
-        <section id="app-widget" class="container app-host-lunch hide">
-            <div class="app-header">
-                <h2>Start a lunch group</h2>
-                <input required id="group-name" type="text" placeholder="Your group name" />
-                <input required id="timepicker-start" class="timepickerStart timeSelect" placeholder="Start Time" jt-timepicker="" time="model.time" time-string="model.timeString" default-time="model.options.defaultTime" time-format="model.options.timeFormat" start-time="model.options.startTime"
-                    min-time="model.options.minTime" max-time="model.options.maxTime" interval="model.options.interval" dynamic="model.options.dynamic" scrollbar="model.options.scrollbar" dropdown="model.options.dropdown">
-                <input required id="timepicker-end" class="timepickerEnd timeSelect" placeholder="End Time" jt-timepicker="" time="model.time" time-string="model.timeString" default-time="model.options.defaultTime" time-format="model.options.timeFormat" start-time="model.options.startTime"
-                    min-time="model.options.minTime" max-time="model.options.maxTime" interval="model.options.interval" dynamic="model.options.dynamic" scrollbar="model.options.scrollbar" dropdown="model.options.dropdown">
-                <select required id="group-size-select">
-                    <option value="" disabled selected hidden>Group Size</option>
-                    <option label="Small (1-2)" value="sm"></option>
-                    <option label="Medium (3-5)" value="md"></option>
-                    <option label="Large (5-8)" value="lg"></option>
-                </select>
-                <button id="profile-btn" type="button" class="timePickerButton btn-red">Create Group</button>
-            </div>
-        </section>
-
-        <section class="container register-login animated fadeInDown faster">
-            <div class="register">
-                <div class="left">
-                    <h1 class="logo"><span class="lunch">LUNCH</span><span class="memo">memo</span></h1>
-                    <p>Lunch Memo makes it easy to organize lunch with your colleagues and other professionals in your area. Spend less time finding lunch spots and scheduling with everyone and simplify your lunch life.
-                    </p>
-                </div>
-                <div class="right widget-register">
-                    <h3><span>Hungry?</span>Create an account below to get started</h3>
-                    <form id="fp-registration">
-                        <label for="name">Name</label>
-                        <input type="text" id="register-name" required minlength=3 />
-
-                        <label for="email">Email</label>
-                        <input type="email" id="register-email" required minlength=3 />
-
-                        <label for="password">Password</label>
-                        <input type="password" id="register-password" required />
-
-                        <div class="form-footer">
-                            <button type="submit" class="btn-red">Register</button>
-                            <span><a href="#" class="go-login">Log in</a></span>
-                        </div>
-
-                        <div class="notice"></div>
-                    </form>
-                </div>
-                <div class="right widget-login hide">
-                    <h3><span>Hungry?</span>Log in below to get connected</h3>
-                    <form id="fp-login">
-                        <label for="email">Email</label>
-                        <input type="email" id="login-email" required minlength=3 />
-
-                        <label for="password">Password</label>
-                        <input type="password" id="login-password" required />
-
-                        <div class="form-footer">
-                            <button type="submit" class="btn-red">Log in</button>
-                            <span><a href="#" class="go-register">Register</a></span>
-                        </div>
-
-                        <div class="notice error"></div>
-                    </form>
-                </div>
-            </div>
-        </section>
-
-        <section class="container groups-wrapper hide">
-        </section>
-    </div>
-    <!-- .app-wrap-->
-
-    <div class="site-background">
-        <div class="bginfo">
-            <a href="#" target="_blank">First Last<br /><span>Title of the photo</span></a>
-        </div>
-    </div>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/timepicker/1.3.5/jquery.timepicker.min.js"></script>
-    <script src="js/api-handler.js"></script>
-    <script src="js/index.js"></script>
-    <script src="js/app.js"></script>
-</body>
-
-</html>
