--- ARCHIVE/src/views/auth/login.hbs
@@ -1,16 +1,0 @@
-<form action="/login" method="POST">
-    <div>
-        <label for="username">Username:</label>
-        <input type="text" name="username">
-    </div>
-    <div>
-        <label for="password">Password:</label>
-        <input type="password" name="password">
-    </div>
-    {{#if message}}
-    <div class="error-message">{{ message }}</div>
-    {{/if}}
-    <div>
-        <input type="submit" value="Log In">
-    </div>
-</form>
