--- ARCHIVE/src/views/auth/signup.hbs
@@ -1,9 +1,0 @@
-<form id="form" action="/signup" method="POST">
-    <label for="username">Username</label><br>
-    <input type="text" name="username" placeholder="Your username">
-
-    <label for="password">Password</label><br>
-    <input type="password" name="password" placeholder="********">
-
-    <button type="submit">Create account</button>
-</form>
