--- ARCHIVE/src/views/layout.hbs
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]><!-->
-<html class="no-js">
-<!--<![endif]-->
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Lunch Memo</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel="apple-touch-icon" sizes="180x180" href="img/icon/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="img/icon/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="img/icon/favicon-16x16.png">
-    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#f5d808">
-    <meta name="apple-mobile-web-app-title" content="LUNCHmemo">
-    <meta name="application-name" content="LUNCHmemo">
-    <meta name="msapplication-TileColor" content="#ff0000">
-    <meta name="theme-color" content="#ffffff">
-    <meta name="google-site-verification" content="PwZNKMZ1g6XsFIqHaovSIzHU24BpRpEUE7xzM2DE4nA" />
-    
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
-
-    <link href="https://fonts.googleapis.com/css?family=Lato:400,700|Rubik:300,300i,700&display=swap" rel="stylesheet">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.7.2/animate.min.css" rel="stylesheet">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.10.2/css/all.min.css" rel="stylesheet">
-    <link href="css/main.css" rel="stylesheet">
-</head>
-
-<body>
-    <!--[if lt IE 7]>
-            <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="#">upgrade your browser</a> to improve your experience.</p>
-        <![endif]-->
-
-    <div class="app-wrap">
-      {{{body}}}
-
-    <div class="site-background">
-        <div class="bginfo">
-            <a href="#" target="_blank">First Last<br /><span>Title of the photo</span></a>
-        </div>
-    </div>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/timepicker/1.3.5/jquery.timepicker.min.js"></script>
-    <script src="js/api-handler.js"></script>
-    <script src="js/index.js"></script>
-    <script src="js/app.js"></script>
-</body>
-
-</html>
