--- ARCHIVE/src/views/not-found.hbs
@@ -1,2 +1,0 @@
-<h1>404 Error</h1>
-<h2>Are you feeling okay?  You seem to have tried to go somewhere that doesn't exist...</h2>
