--- ARCHIVE/src/views/users/login.hbs
@@ -1,12 +1,0 @@
-<h1>Login</h1>
-
-<form action="/login" method="POST">
-  <p>Username</p>
-  <input type="text" name="username">
-
-  <p>Password</p>
-  <input type="password" name="password">
-
-  <button type="submit">Log In</button>
-</form>
-{{> alert }}
