--- ARCHIVE/src/views/users/messages.hbs
@@ -1,2 +1,0 @@
-<h2>Welcome {{user.username}}</h2>
-<p>Your messages would be here if you had any...</p>
