--- ARCHIVE/src/views/users/signup.hbs
@@ -1,12 +1,0 @@
-<h1>Signup</h1>
-
-<form action="/signup" method="POST">
-  <p>Username</p>
-  <input type="text" name="username">
-
-  <p>Password</p>
-  <input type="password" name="password">
-
-  <button type="submit">Save</button>
-</form>
-{{> alert }}
