+++ app/services/api/users.ts
@@ -1,5 +1,7 @@
-export const getProfiles = async (uids: string) => {
-  const response = await fetch(`/api/users/getProfiles`, {
+import { GetUserProfilesResponse } from '@typing/api';
+
+export const getProfiles = async (uids: string[]): Promise<GetUserProfilesResponse> => {
+  const response = await fetch(`/api/users/profiles`, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({ uids }) //
